+++ windows-terminal-quake/Logging.cs
@@ -11,20 +11,16 @@ namespace WindowsTerminalQuake
 		{
 			var here = Path.GetDirectoryName(new Uri(typeof(Logging).Assembly.Location).LocalPath);
 
-			var builder = new LoggerConfiguration()
-				.MinimumLevel.Is(Serilog.Events.LogEventLevel.Information)
-			;
-
-			builder
+			Log.Logger = new LoggerConfiguration()
 				.MinimumLevel.Is(QSettings.Instance.LogLevel)
 				.WriteTo.File(
 					path: Path.Combine(here, "logs/.txt"),
 					fileSizeLimitBytes: 10_000_000,
 					rollingInterval: RollingInterval.Day,
 					retainedFileCountLimit: 3
-				);
-
-			Log.Logger = builder.CreateLogger();
+				)
+				.CreateLogger()
+			;
 
 			Log.Information("Windows Terminal Quake started");
 		}
