+++ windows-terminal-quake/Native/HotKeyManager.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Threading;
 using System.Windows.Forms;
+using WindowsTerminalQuake.Settings;
 
 namespace WindowsTerminalQuake.Native
 {
@@ -99,15 +100,5 @@ namespace WindowsTerminalQuake.Native
 			Key = (Keys)((param & 0xffff0000) >> 16);
 			Modifiers = (KeyModifiers)(param & 0x0000ffff);
 		}
-	}
-
-	[Flags]
-	public enum KeyModifiers
-	{
-		Alt = 1,
-		Control = 2,
-		Shift = 4,
-		Windows = 8,
-		NoRepeat = 0x4000
 	}
 }
