+++ windows-terminal-quake/Settings/HorizontalAlign.cs
@@ -1,0 +1,10 @@
+﻿namespace WindowsTerminalQuake.Settings
+{
+	public enum HorizontalAlign
+	{
+		Center = 0,
+
+		Left,
+		Right
+	}
+}
