+++ windows-terminal-quake/Settings/Hotkey.cs
@@ -1,0 +1,12 @@
+﻿using System.Windows.Forms;
+using WindowsTerminalQuake.Native;
+
+namespace WindowsTerminalQuake.Settings
+{
+	public class Hotkey
+	{
+		public Keys Key { get; set; }
+
+		public KeyModifiers Modifiers { get; set; }
+	}
+}
